+++ kits/rokt/rokt/src/main/kotlin/com/mparticle/kits/RoktEmbeddedView.kt
@@ -1,4 +1,4 @@
-package com.mparticle.rokt
+package com.mparticle.kits
 
 import android.content.Context
 import android.util.AttributeSet
