+++ kits/rokt/rokt/src/main/kotlin/com/mparticle/kits/RoktLayoutDimensionCallBack.kt
@@ -1,4 +1,4 @@
-package com.mparticle.rokt
+package com.mparticle.kits
 
 interface RoktLayoutDimensionCallBack {
     fun onHeightChanged(height: Int)
